Fix #110. Guards/Actions of transitions with event<_> are considered.
Now, it is possible to create a transition-table that has multiple
transitions that match any event (via "event<_>") but still can use
guards and actions to distinguish between them.

However, the events that should be matched by such transitions still
must be used (verbatim) in some other (possibly unreachable) transition
of the transition-table or they would be considered as 'unexpected'
events and not being considered by such "catch-all" transitions.